Green: Refactored MatchFlightsAndHotels and Passed the Test.

+++ HolidaySearch/SearchModels/MatchFlightsAndHotels.cs
@@ -6,5 +6,5 @@ public class MatchFlightsAndHotels
 {
     public List<Flights> FlightsList { get; set; }
     public List<Hotels> HotelsList { get; set; }
-    public List<int> ListOfTotalPrices { get; set; }
+    public List<decimal> ListOfTotalPrices { get; set; }
 }
